requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary
Pillow
django-filter
drf-yasg
python-dotenv
